# src/main/resources/api.yml
openapi: 3.0.0
info:
  title: Gestión de Transacciones
  version: 1.0.0
  description: API para gestionar transacciones bancarias

paths:
  /transacciones/deposito:
    post:
      tags:
        - transacciones
      summary: Registrar un depósito
      description: Permite registrar un depósito en una cuenta
      operationId: registrarDeposito
      requestBody:
        description: Información del depósito
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Deposito'
      responses:
        '200':
          description: Depósito registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '400':
          description: Datos inválidos

  /transacciones/retiro:
    post:
      tags:
        - transacciones
      summary: Registrar un retiro
      description: Permite registrar un retiro de una cuenta
      operationId: registrarRetiro
      requestBody:
        description: Información del retiro
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Retiro'
      responses:
        '200':
          description: Retiro registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '400':
          description: Datos inválidos
        '422':
          description: Saldo insuficiente

  /transacciones/transferencia:
    post:
      tags:
        - transacciones
      summary: Registrar una transferencia
      description: Permite registrar una transferencia entre cuentas
      operationId: registrarTransferencia
      requestBody:
        description: Información de la transferencia
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transferencia'
      responses:
        '200':
          description: Transferencia registrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '400':
          description: Datos inválidos
        '422':
          description: Saldo insuficiente

  /transacciones/historial:
    get:
      tags:
        - transacciones
      summary: Consultar historial de transacciones
      description: Obtiene el historial de transacciones de una cuenta
      operationId: obtenerHistorial
      parameters:
        - name: cuentaId
          in: query
          description: ID de la cuenta para consultar el historial
          required: true
          schema:
            type: integer
        - name: fechaInicio
          in: query
          description: Fecha de inicio del rango a consultar (opcional)
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          description: Fecha de fin del rango a consultar (opcional)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Historial obtenido exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaccion'
        '400':
          description: Datos inválidos

components:
  schemas:
    Deposito:
      type: object
      required:
        - cuentaId
        - monto
        - fecha
      properties:
        cuentaId:
          type: integer
          description: ID de la cuenta a depositar
        monto:
          type: number
          format: float
          description: Monto a depositar
        fecha:
          type: string
          format: date-time
          description: Fecha del depósito

    Retiro:
      type: object
      required:
        - cuentaId
        - monto
        - fecha
      properties:
        cuentaId:
          type: integer
          description: ID de la cuenta a retirar
        monto:
          type: number
          format: float
          description: Monto a retirar
        fecha:
          type: string
          format: date-time
          description: Fecha del retiro

    Transferencia:
      type: object
      required:
        - cuentaOrigenId
        - cuentaDestinoId
        - monto
        - fecha
      properties:
        cuentaOrigenId:
          type: integer
          description: ID de la cuenta de origen
        cuentaDestinoId:
          type: integer
          description: ID de la cuenta de destino
        monto:
          type: number
          format: float
          description: Monto a transferir
        fecha:
          type: string
          format: date-time
          description: Fecha de la transferencia

    Transaccion:
      type: object
      required:
        - tipo
        - monto
        - fecha
      properties:
        tipo:
          type: string
          enum:
            - depósito
            - retiro
            - transferencia
          description: Tipo de la transacción
        monto:
          type: number
          format: float
          description: Monto de la transacción
        fecha:
          type: string
          format: date-time
          description: Fecha de la transacción
        cuentaOrigenId:
          type: integer
          description: ID de la cuenta de origen (si aplica)
        cuentaDestinoId:
          type: integer
          description: ID de la cuenta de destino (si aplica)

    Respuesta:
      type: object
      properties:
        codigo:
          type: integer
          format: int32
        mensaje:
          type: string